+++ infrastructure/management/src/main/java/com/bytex/snamp/management/shell/DeleteOperationCommand.java
@@ -4,35 +4,35 @@ import com.bytex.snamp.SpecialUse;
 import com.bytex.snamp.configuration.AgentConfiguration;
 import org.apache.karaf.shell.commands.Argument;
 import org.apache.karaf.shell.commands.Command;
-import static com.bytex.snamp.configuration.AgentConfiguration.ManagedResourceConfiguration.AttributeConfiguration;
+import static com.bytex.snamp.configuration.AgentConfiguration.ManagedResourceConfiguration.OperationConfiguration;
 
 /**
- * Deletes attribute from managed resource.
+ * Deletes operation from managed resource.
  * @author Roman Sakno
  * @version 1.0
  * @since 1.0
  */
 @Command(scope = SnampShellCommand.SCOPE,
-    name = "delete-attribute",
-    description = "Delete attribute from managed resource")
-public final class DeleteAttributeCommand extends ConfigurationCommand {
+    name = "delete-operation",
+    description = "Delete operation from managed resource")
+public final class DeleteOperationCommand extends ConfigurationCommand {
     @Argument(index = 0, name = "resourceName", required = true, description = "Name of the managed resource to modify")
     @SpecialUse
     private String resourceName = "";
 
-    @Argument(index = 1, name = "userDefinedName", required = true, description = "User-defined name of the attribute to remove")
+    @Argument(index = 1, name = "userDefinedName", required = true, description = "User-defined name of the operation to remove")
     @SpecialUse
     private String userDefinedName = "";
 
     @Override
     boolean doExecute(final AgentConfiguration configuration, final StringBuilder output) {
         if(configuration.getManagedResources().containsKey(resourceName))
-            if(configuration.getManagedResources().get(resourceName).getElements(AttributeConfiguration.class).remove(userDefinedName) == null){
-                output.append("Attribute doesn't exist");
+            if(configuration.getManagedResources().get(resourceName).getElements(OperationConfiguration.class).remove(userDefinedName) == null){
+                output.append("Operation doesn't exist");
                 return false;
             }
             else {
-                output.append("Attribute deleted successfully");
+                output.append("Operation deleted successfully");
                 return true;
             }
         else {